Changed StatementCode to a ushort. Added save and load statements (but not code to parse them).

+++ LSN/LsnCore/Statements/AssignmentStatement.cs
@@ -30,7 +30,7 @@ namespace LsnCore.Statements
 
 		internal override void Serialize(BinaryDataWriter writer, ResourceSerializer resourceSerializer)
 		{
-			writer.Write((byte)StatementCode.AssignVariable);
+			writer.Write(StatementCode.AssignVariable);
 			writer.Write((ushort)Index);
 			Value.Serialize(writer, resourceSerializer);
 		}
