+++ LSN/LsnCore/Statements/DisplayChoicesStatement.cs
@@ -21,7 +21,7 @@ namespace LsnCore.Statements
 
 		internal override void Serialize(BinaryDataWriter writer, ResourceSerializer resourceSerializer)
 		{
-			writer.Write((byte)StatementCode.DisplayChoices);
+			writer.Write(StatementCode.DisplayChoices);
 		}
 	}
 }
