+++ LSN/LsnCore/Statements/ExpressionStatement.cs
@@ -33,7 +33,7 @@ namespace LsnCore.Statements
 
 		internal override void Serialize(BinaryDataWriter writer, ResourceSerializer resourceSerializer)
 		{
-			writer.Write((byte)StatementCode.EvaluateExpression);
+			writer.Write(StatementCode.EvaluateExpression);
 			Expression.Serialize(writer, resourceSerializer);
 		}
 	}
