+++ LSN/LsnCore/Statements/GiveStatement.cs
@@ -37,7 +37,7 @@ namespace LsnCore.Statements
 
 		internal override void Serialize(BinaryDataWriter writer, ResourceSerializer resourceSerializer)
 		{
-			writer.Write((byte)StatementCode.GiveItem);
+			writer.Write(StatementCode.GiveItem);
 			Amount.Serialize(writer, resourceSerializer);
 			Id.Serialize(writer, resourceSerializer);
 			Receiver.Serialize(writer, resourceSerializer);
@@ -65,7 +65,7 @@ namespace LsnCore.Statements
 
 		internal override void Serialize(BinaryDataWriter writer, ResourceSerializer resourceSerializer)
 		{
-			writer.Write((byte)StatementCode.GiveGold);
+			writer.Write(StatementCode.GiveGold);
 			Amount.Serialize(writer, resourceSerializer);
 			Receiver.Serialize(writer, resourceSerializer);
 		}
